set show me value in ion-select when set in the store
