+++ src/modules/tests/vehicle-checks/vehicle-checks.selector.ts
@@ -1,5 +1,6 @@
 import { VehicleChecks } from '@dvsa/mes-test-schema/categories/B';
 import { default as tellMeQuestions } from '../../../providers/question/tell-me-question.constants';
+import { default as showMeQuestions } from '../../../providers/question/show-me-question.constants';
 
 export const isTellMeQuestionSelected = (state: VehicleChecks) => state.tellMeQuestionCode !== undefined;
 export const isTellMeQuestionCorrect = (state: VehicleChecks) => state.tellMeQuestionOutcome === 'P';
@@ -12,3 +13,6 @@ export const getSelectedTellMeQuestionText = (state: VehicleChecks) => {
   }
   return `${state.tellMeQuestionCode} - ${tellMeQuestionText.tellMeQuestionShortName}`;
 };
+
+export const getShowMeQuestion = (state: VehicleChecks) =>
+  showMeQuestions.find(question => question.showMeQuestionCode === state.showMeQuestionCode);
