+++ src/pages/office/office.html
@@ -182,14 +182,14 @@
               <ion-row class="spacing-row"></ion-row>
               <ion-row>
                 <ion-select (ionChange)="showMeQuestionChanged($event)" placeholder="Select a question"
-                  id="show-me-selector" value="show-me-question" okText="Submit" cancelText="Cancel">
+                  id="show-me-selector"  [(ngModel)]="showMeQuestion" okText="Submit" cancelText="Cancel" formControlName="showMeQuestionCtrl">
                   <ion-option [value]="showMeQuestion" *ngFor="let showMeQuestion of showMeQuestions">
                     {{showMeQuestion.showMeQuestionCode}} - {{showMeQuestion.showMeQuestionShortName}}
                   </ion-option>
                 </ion-select>
               </ion-row>
-              <ion-row class="validation-message-row" align-items-center>
-                <div class="validation-text">
+              <ion-row class="validation-message-row" align-it  ems-center>
+                <div class="validation-text" [class.ng-invalid]="isCtrlDirtyAndInvalid('showMeQuestionCtrl')">
                   Enter the show me question
                 </div>
               </ion-row>
